requirements.txt
streamlit
sentence-transformers
faiss-cpu
pymupdf
groq
numpy
python-dotenv